refactor(baseball): add modules package
modules 패키지를 추가하여 클래스들을 분리하였다.
