+++ src/main/java/baseball/Computer.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.modules.RandomNumberPicker;
+
 public class Computer {
 
 	private RandomNumberPicker randomNumberPicker = new RandomNumberPicker();
