+++ src/main/java/baseball/modules/RandomNumberPicker.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.modules;
 
 import java.util.HashSet;
